move QgsCompoundCurve tests from QgsGeometry
